+++ src/core/managers/internal/SignTransactionsStateManager/types/signTransactionsPanel.types.ts
@@ -39,6 +39,7 @@ export interface ISignTransactionsPanelCommonData {
   isEditable?: boolean;
   highlight?: string | null;
   scCall?: string | null;
+  providerName?: string;
 }
 
 export interface ISignTransactionsPanelData {